updated uid in db to bigint from integer

+++ db/migrate/20180707145608_create_users.rb
@@ -1,7 +1,7 @@
 class CreateUsers < ActiveRecord::Migration[5.2]
   def change
     create_table :users do |t|
-      t.integer  :uid
+      t.bigint  :uid
       t.string :name
       t.string :image
       t.string   :first_name
